Use BitSet.length() instead of BitSet.size()
The size() method on bitSet returns the size of the underlying native
data structure representing the bit set, which is always rounded up to
the nearest multiple of 64. This causes the declared CIDSet <= font
CIDset subset check to become overly strict if the number of glyphs in
the CIDSet is not an exact multiple of 64.